% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xiaowei.R
\name{func_credit_use}
\alias{func_credit_use}
\title{xiaowei pass and use}
\usage{
func_credit_use(qudao)
}
\arguments{
\item{qudao}{}
}
\value{
dataframe
}
\description{
xiaowei pass and use
}
\examples{

}
