% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xiaowei.R
\name{getMifiJdbcQuery}
\alias{getMifiJdbcQuery}
\title{The function to connect to datdabases}
\usage{
getMifiJdbcQuery(sql = NULL, database = "mifi", queue = "loanba")
}
\arguments{
\item{queue}{}
}
\value{
results of sql
}
\description{
The function to connect to datdabases
}
